Finished the basic implementation of finding the shortest route and the fasted route. feat: the basic implementation of finding the shortest route and the fasted route. fix: the wrong memory delivery.

+++ dataStructure.h
@@ -34,7 +34,7 @@ struct geom
 struct edge
 {
     long id;
-    long to, nxt;
+    int to, nxt;
     double length;
 };
